Code cleanup, reduction, and documentation. No algorithmic changes.
No longer compiling using -pedantic (such that we may print uint64_t)
Added the compilation flag -Wshadow
Added const qualifier to many of the lookup tables
Added a genAllLegalMoves routine (used to create root move list in search)
Added a great deal of documenation into Magics.c
Added function documentation into History.c
Added u and ull to unsigned and uint64_t data types (most, if not all)
Added some documentation into search.c, more to come
Removed Tempo constants, and replace with const int Tempo[PHASE_NB]
Removed stray calculations for rank and file (Now always using the Macros)
Removed precomputed Magic Index Tables (Now computed at start up)
Removed an unused BitMask from Mask.c (PawnAdvanceMasks)
Removed unused code from UCI.c, and generally cleaned it up
Converted some counts (PERFT, nodes) to uint64_t
Credited Rebel Chess and CPW for the Magic code and numbers
Changed some Magic Numbers using the best of table on CPW
General cleanup of spacing, character limits, ect

+++ src/board.h
@@ -23,7 +23,7 @@
 
 void initalizeBoard(Board * board, char * fen);
 void printBoard(Board * board);
-int perft(Board * board, int depth);
+uint64_t perft(Board * board, int depth);
 void runBenchmark(int depth);
 
 #endif
